moving solvers to a separate folder
